+++ libcoveoc4ids/__main__.py
@@ -16,12 +16,7 @@ def main():
 
     cove_temp_folder = tempfile.mkdtemp(prefix='lib-cove-oc4ids-cli-', dir=tempfile.gettempdir())
     try:
-        result = oc4ids_json_output(
-            cove_temp_folder,
-            args.filename,
-            file_type='json',
-            cache=False,
-        )
+        result = oc4ids_json_output(cove_temp_folder, args.filename, file_type='json')
     finally:
         shutil.rmtree(cove_temp_folder)