+++ libcoveoc4ids/api.py
@@ -10,40 +10,37 @@ except ImportError:
     import json  # noqa: F401
 
 
-class APIException(Exception):
-    pass
-
-
-def oc4ids_json_output(output_dir, file, file_type=None, json_data=None,
-                       lib_cove_oc4ids_config=None, cache=True):
-
+def oc4ids_json_output(
+    output_dir: str = "",
+    file=None,  # : str | None
+    file_type=None,  # : str | None
+    json_data=None,  # : dict | None
+    lib_cove_oc4ids_config=None,  # : LibCoveOC4IDSConfig | None
+):
     if not lib_cove_oc4ids_config:
         lib_cove_oc4ids_config = LibCoveOC4IDSConfig()
 
     if not file_type:
         file_type = get_file_type(file)
+
     context = {"file_type": file_type}
 
-    if file_type == 'json':
+    if file_type == "json":
         if not json_data:
-            with open(file, "rb") as fp:
-                try:
-                    json_data = json.loads(fp.read())
-                except ValueError:
-                    raise APIException('The file looks like invalid json')
+            with open(file, "rb") as f:
+                json_data = json.loads(f.read())
 
-        schema_oc4ids = SchemaOC4IDS(lib_cove_oc4ids_config=lib_cove_oc4ids_config)
-
+        schema_obj = SchemaOC4IDS(lib_cove_oc4ids_config)
     else:
-
-        raise Exception("JSON only for now, sorry!")
+        raise NotImplementedError
 
     context = common_checks_oc4ids(
         context,
         output_dir,
         json_data,
-        schema_oc4ids,
-        lib_cove_oc4ids_config=lib_cove_oc4ids_config,
-        cache=cache)
+        schema_obj,
+        # common_checks_context(cache=True) caches the results to a file, which is not needed in API context.
+        cache=False,
+    )
 
     return context